credential_injector: move integration tests to the extensions package
