+++ test/extensions/http/credential_injector/generic/BUILD
@@ -33,8 +33,10 @@ envoy_extension_cc_test(
     rbe_pool = "6gig",
     deps = [
         "//source/extensions/filters/http/credential_injector:config",
+        "//source/extensions/filters/network/tcp_proxy:config",
         "//source/extensions/http/injected_credentials/generic:config",
         "//test/integration:tcp_tunneling_integration_lib",
-        "//test/test_common:utility_lib",
+        "@envoy_api//envoy/extensions/filters/http/credential_injector/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/filters/network/tcp_proxy/v3:pkg_cc_proto",
     ],
 )
