+++ tools/code_format/config.yaml
@@ -445,12 +445,9 @@ visibility_excludes:
 - source/extensions/clusters/dns/
 - source/extensions/early_data/BUILD
 - source/extensions/filters/http/buffer/BUILD
-- source/extensions/filters/http/credential_injector/BUILD
 - source/extensions/filters/network/common/BUILD
 - source/extensions/filters/network/generic_proxy/interface/BUILD
 - source/extensions/http/header_validators/envoy_default/BUILD
-- source/extensions/http/injected_credentials/common/BUILD
-- source/extensions/http/injected_credentials/generic/BUILD
 - source/extensions/transport_sockets/common/BUILD
 - source/extensions/transport_sockets/tap/BUILD
 - source/extensions/udp_packet_writer/default/BUILD
